requirements.txt
streamlit==1.40.0
python-dotenv==1.0.1
pydantic==2.10.3
google-generativeai==0.8.3